requirements.txt
torch
torchvision
torchaudio
langchain==0.0.184
langchain-community==0.0.10
PyPDF2==3.0.1
python-dotenv==1.2.1
streamlit==1.52.2

faiss-cpu==1.13.1
altair==6.0.0

# For Instructor Embeddings
huggingface-hub==0.36.0
sentence-transformers==5.2.0
transformers==4.57.3
accelerate==1.12.0
bitsandbytes==0.49.0
